Fixed OX-3905 in OpenX 2.7.31-beta.
Manually changing account while in statistics screens will now re-direct user to an appropriate page in the statistics system if they do not have permission to view the existing page after switching. Manually entering a URL that the user/account has no permission to view also now correctly shows an error message that permission is denied, and allows the user to log in as a different user.



git-svn-id: https://svn.openx.org/openx/trunk@34013 69d48449-a624-0410-ae76-85ddf3b22ea5